refactor(listitemrow.tsx): moved to root components/ folder
