+++ app/index.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { ActivityIndicator, FlatList } from 'react-native';
 
-import ListItemRow from '~/components/app/ListItemRow';
+import ListItemRow from '~/components/ListItemRow';
 import { ThemedText } from '~/components/ThemedText';
 import { ListItem, useGetListsQuery } from '~/lib/supabase/supabaseAPI';
 import { getItemLayout } from '~/lib/utils';
